question_app/requirements.txt
astroid==2.2.5
autopep8==1.4.4
Babel==2.7.0
beautifulsoup4==4.8.1
bs4==0.0.1
cairocffi==1.1.0
CairoSVG==2.5.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cssselect2==0.2.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.2.18
django-appconf==1.0.4
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.5.0
django-countries==7.2.1
django-crispy-forms==1.7.2
django-currencies==0.9.0
django-heroku==0.3.1
django-import-export==2.3.0
django-js-asset==1.2.2
django-model-utils==3.2.0
django-money==0.15.1
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-notification==1.3.3
django-notifications-hq==1.6.0
django-session-timeout==0.1.0
django-user-accounts==3.0.2
et-xmlfile==1.0.1
fontawesome-free==5.15.2
future==0.18.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
isort==4.3.21
jdcal==1.4.1
jsonfield==3.1.0
lazy-object-proxy==1.4.2
MarkupPy==1.14
mccabe==0.6.1
numpy==1.20.3
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.1.1
progressbar2==3.47.0
psycopg2==2.8.3
py-moneyed==0.8.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-utils==2.3.0
pytz==2019.3
PyYAML==5.4
pyzk==0.9
reportlab==3.5.32
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
swapper==1.1.1
tablib==2.0.0
tinycss2==1.0.2
tqdm==4.36.1
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.6
WeasyPrint==51
webencodings==0.5.1
whitenoise==3.2
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
